refactor(dummy-plugin): Moved dummy plugin files into dev.avatcher.cinnamon.dummy package
