+++ dummy-plugin/src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
 name: CDummy
 version: '${version}'
-main: dev.avatcher.cdummy.CDummy
+main: dev.avatcher.cinnamon.dummy.CDummy
 api-version: '1.20'
 load: STARTUP
 depend: [ Cinnamon ]
